prism: split internal/db/db.go into per-concern files (D1.OPEN + D2.DB)
Decompose db.Open() into helper functions (openAndConfigure,
seedSchemaVersionIfEmpty, runMigrations, migrateV*toV*) and split the
3892-line db.go into 9 files within package db:

- db.go        — DB struct, Open/Close, schema, migrations
- types.go     — all struct/type definitions
- status.go    — agent_status CRUD operations
- events.go    — event write/query operations
- bus.go       — bus message operations
- sessions.go  — session table operations
- groups.go    — session group operations
- mergequeue.go — pending merge queue operations
- maintenance.go — Prune()

No exported symbols changed; all callers compile without modification.

Closes #1150
