+++ modules/programs/prism/prism/internal/db/status.go
@@ -1,0 +1,856 @@
+package db
+
+import (
+	"database/sql"
+	"errors"
+	"fmt"
+	"net"
+	"os"
+	"strings"
+	"time"
+
+	"github.com/prismatic-koi/prism/internal/agent"
+)
+
+// currentStateOf looks up the current agent state for sessionName. Returns
+// ("", nil) when no row exists (fresh insert path — caller should skip
+// transition validation).
+func (d *DB) currentStateOf(sessionName string) (agent.AgentState, error) {
+	var state string
+	err := d.conn.QueryRow(
+		"SELECT state FROM agent_status WHERE session_name = ?", sessionName,
+	).Scan(&state)
+	if errors.Is(err, sql.ErrNoRows) {
+		return "", nil
+	}
+	if err != nil {
+		return "", fmt.Errorf("db: read current state: %w", err)
+	}
+	return agent.AgentState(state), nil
+}
+
+// checkTransition reads the current state for sessionName and validates that
+// transitioning to toState is permitted. When the transition is invalid it
+// logs a warning to stderr (including session name, from→to pair, and caller
+// context) and returns — it does not return an error, so callers are never
+// blocked by an invalid transition.
+//
+// Same-state "transitions" (fromState == toState) are always silently skipped:
+// they represent metadata-only upserts (title, harness_session_id, model_id,
+// last_seen) where the state value does not actually change, so there is
+// nothing to validate.
+//
+// Callers pass a short context string (e.g. "UpsertStatus", "pane-died") that
+// is included in the log line to help locate the call site.
+func (d *DB) checkTransition(sessionName string, toState agent.AgentState, callerCtx string) {
+	fromState, err := d.currentStateOf(sessionName)
+	if err != nil {
+		// Non-fatal: if we can't read the current state we skip validation.
+		fmt.Fprintf(os.Stderr, "[prism] %s: could not read current state for %q: %v\n",
+			callerCtx, sessionName, err)
+		return
+	}
+	if fromState == "" {
+		// No prior row — fresh insert; no transition to validate.
+		return
+	}
+	if fromState == toState {
+		// Same-state update — metadata-only refresh, nothing to validate.
+		return
+	}
+	if err := agent.Transition(fromState, toState); err != nil {
+		fmt.Fprintf(os.Stderr, "[prism] %s: invalid transition for session %q: %v\n",
+			callerCtx, sessionName, err)
+	}
+}
+
+// UpsertStatus inserts or updates the agent_status row for sessionName.
+// repo and worktree are always overwritten on conflict. title, harnessSessionID,
+// agentName, and modelID are updated only when non-nil (COALESCE).
+func (d *DB) UpsertStatus(sessionName, repo, worktree, state string, title *string, harnessSessionID *string) error {
+	return d.UpsertStatusWithAgent(sessionName, repo, worktree, state, title, harnessSessionID, nil, nil)
+}
+
+// UpsertStatusWithAgent is like UpsertStatus but also accepts agentName and
+// modelID, which are written to agent_status.agent_name and agent_status.model_id
+// using COALESCE (only overwriting when non-nil). root_agent_name and root_model_id
+// are NOT touched by this method — use UpsertStatusWithRootAgent for session creation.
+func (d *DB) UpsertStatusWithAgent(sessionName, repo, worktree, state string, title *string, harnessSessionID *string, agentName *string, modelID *string) error {
+	d.checkTransition(sessionName, agent.AgentState(state), "UpsertStatusWithAgent")
+	now := time.Now().UnixMilli()
+	const q = `
+INSERT INTO agent_status (session_name, repo, worktree, state, title, agent_name, model_id, last_seen, harness, harness_session_id)
+VALUES (?, ?, ?, ?, ?, ?, ?, ?, 'opencode', ?)
+ON CONFLICT(session_name) DO UPDATE SET
+  state              = excluded.state,
+  repo               = excluded.repo,
+  worktree           = excluded.worktree,
+  title              = COALESCE(excluded.title, title),
+  agent_name         = COALESCE(excluded.agent_name, agent_name),
+  model_id           = COALESCE(excluded.model_id, model_id),
+  last_seen          = excluded.last_seen,
+  harness            = 'opencode',
+  harness_session_id = COALESCE(excluded.harness_session_id, harness_session_id)`
+	_, err := d.conn.Exec(q, sessionName, repo, worktree, state, title, agentName, modelID, now, harnessSessionID)
+	if err != nil {
+		return fmt.Errorf("db: upsert status: %w", err)
+	}
+	return nil
+}
+
+// UpdateRootModelID unconditionally sets root_model_id for sessionName to the
+// given model value. Unlike UpsertStatusWithRootAgent (which falls back to the
+// existing value when the incoming value is nil), this method always overwrites,
+// allowing the current session's model to replace a stale value from a prior
+// session.
+//
+// It is a no-op when no row exists for sessionName (returns nil).
+// Called by the sidecar when a completed assistant message from the root agent
+// reveals the current model, so that coordinator notifications always reflect
+// the live model configuration.
+func (d *DB) UpdateRootModelID(sessionName, modelID string) error {
+	_, err := d.conn.Exec(
+		"UPDATE agent_status SET root_model_id = ? WHERE session_name = ?",
+		modelID, sessionName,
+	)
+	if err != nil {
+		return fmt.Errorf("db: update root_model_id: %w", err)
+	}
+	return nil
+}
+
+// UpsertStatusSeedRootAgentName is like UpsertStatus but also writes
+// rootAgentName to root_agent_name when it is non-empty. On conflict (update),
+// root_agent_name is written via COALESCE — the existing value is preserved if
+// the incoming rootAgentName is empty.
+//
+// This is the spawn-time seeding path: called when we know the agent role at
+// session creation time (before the sidecar's first upsertState() call fires),
+// so that the DB row has a non-NULL root_agent_name from the first moment.
+// The sidecar will later write the same value idempotently via
+// UpsertStatusWithRootAgent (COALESCE preserves the already-set value).
+func (d *DB) UpsertStatusSeedRootAgentName(sessionName, repo, worktree, state string, title *string, harnessSessionID *string, rootAgentName string) error {
+	d.checkTransition(sessionName, agent.AgentState(state), "UpsertStatusSeedRootAgentName")
+	now := time.Now().UnixMilli()
+	// When rootAgentName is empty, fall back to leaving root_agent_name as-is
+	// (COALESCE with NULL excluded value preserves existing). When non-empty,
+	// write it, but still use COALESCE so a later sidecar write of the same
+	// value doesn't produce a spurious update.
+	var rootAgentNamePtr *string
+	if rootAgentName != "" {
+		rootAgentNamePtr = &rootAgentName
+	}
+	const q = `
+INSERT INTO agent_status (session_name, repo, worktree, state, title, root_agent_name, last_seen, harness, harness_session_id)
+VALUES (?, ?, ?, ?, ?, ?, ?, 'opencode', ?)
+ON CONFLICT(session_name) DO UPDATE SET
+  state              = excluded.state,
+  repo               = excluded.repo,
+  worktree           = excluded.worktree,
+  title              = COALESCE(excluded.title, title),
+  root_agent_name    = COALESCE(excluded.root_agent_name, root_agent_name),
+  last_seen          = excluded.last_seen,
+  harness            = 'opencode',
+  harness_session_id = COALESCE(excluded.harness_session_id, harness_session_id)`
+	_, err := d.conn.Exec(q, sessionName, repo, worktree, state, title, rootAgentNamePtr, now, harnessSessionID)
+	if err != nil {
+		return fmt.Errorf("db: upsert status seed root agent name: %w", err)
+	}
+	return nil
+}
+
+// UpsertStatusWithRootAgent is like UpsertStatusWithAgent but also writes
+// root_agent_name and root_model_id. On conflict (update), root_agent_name and
+// root_model_id prefer the incoming (excluded) value via COALESCE — the sidecar
+// is authoritative and can correct a stale or wrong value on every state update.
+//
+// The Go sidecar is the authoritative source of root_agent_name: it calls this
+// method on every state transition when Config.AgentRole is set. Because the
+// sidecar value takes precedence, a row written with the wrong root_agent_name
+// (e.g. from a legacy or race-condition write) is corrected on the very next
+// sidecar call. The TypeScript plugin (prism-hooks.ts) does not write
+// root_agent_name.
+func (d *DB) UpsertStatusWithRootAgent(sessionName, repo, worktree, state string, title *string, harnessSessionID *string, agentName *string, modelID *string) error {
+	d.checkTransition(sessionName, agent.AgentState(state), "UpsertStatusWithRootAgent")
+	now := time.Now().UnixMilli()
+	const q = `
+INSERT INTO agent_status (session_name, repo, worktree, state, title, agent_name, model_id, root_agent_name, root_model_id, last_seen, harness, harness_session_id)
+VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, 'opencode', ?)
+ON CONFLICT(session_name) DO UPDATE SET
+  state              = excluded.state,
+  repo               = excluded.repo,
+  worktree           = excluded.worktree,
+  title              = COALESCE(excluded.title, title),
+  agent_name         = COALESCE(excluded.agent_name, agent_name),
+  model_id           = COALESCE(excluded.model_id, model_id),
+  root_agent_name    = COALESCE(excluded.root_agent_name, root_agent_name),
+  root_model_id      = COALESCE(excluded.root_model_id, root_model_id),
+  last_seen          = excluded.last_seen,
+  harness            = 'opencode',
+  harness_session_id = COALESCE(excluded.harness_session_id, harness_session_id)`
+	_, err := d.conn.Exec(q, sessionName, repo, worktree, state, title, agentName, modelID, agentName, modelID, now, harnessSessionID)
+	if err != nil {
+		return fmt.Errorf("db: upsert status with root agent: %w", err)
+	}
+	return nil
+}
+
+// UpsertStatusIfNotTerminal upserts the state for sessionName only when the
+// current state is not already a terminal state (error, finished, interrupted,
+// or deleted) and the session has not yet been ended (ended_at IS NULL). Returns
+// (true, nil) if the update was applied, (false, nil) if the session was
+// already in a terminal state, has been ended, or did not exist, or
+// (false, err) on a database error.
+//
+// This is used by the pane-died hook to transition active sessions to
+// "interrupted" without clobbering a clean "finished" or an "error" state
+// written directly by the sidecar on startup failure, and without acting on
+// sessions that have already been ended by cleanup.
+func (d *DB) UpsertStatusIfNotTerminal(sessionName, state string) (bool, error) {
+	// Snapshot the current state before the write so that the advisory
+	// transition check (below) sees the from-state rather than the newly
+	// written to-state.
+	fromState, _ := d.currentStateOf(sessionName)
+	now := time.Now().UnixMilli()
+	const q = `
+UPDATE agent_status
+SET state = ?, last_seen = ?
+WHERE session_name = ?
+  AND ended_at IS NULL
+  AND state NOT IN ('error', 'finished', 'interrupted', 'deleted')`
+	res, err := d.conn.Exec(q, state, now, sessionName)
+	if err != nil {
+		return false, fmt.Errorf("db: upsert status if not terminal: %w", err)
+	}
+	n, err := res.RowsAffected()
+	if err != nil {
+		return false, fmt.Errorf("db: upsert status if not terminal: rows affected: %w", err)
+	}
+	// Validate only when the write was actually applied (n > 0) and we have a
+	// prior state to validate from. When the SQL WHERE clause suppresses the
+	// write (session already terminal), there is no transition to check.
+	if n > 0 && fromState != "" {
+		if terr := agent.Transition(fromState, agent.AgentState(state)); terr != nil {
+			fmt.Fprintf(os.Stderr, "[prism] UpsertStatusIfNotTerminal: invalid transition for session %q: %v\n",
+				sessionName, terr)
+		}
+	}
+	return n > 0, nil
+}
+
+// UpsertStatusInterruptedOverrideFinished transitions the session to
+// "interrupted", allowing it to override a "finished" state in addition to the
+// active states that UpsertStatusIfNotTerminal covers.  This is used by the
+// pane-died hook when the pane exited with a non-zero exit code: a non-zero
+// exit means the process was killed or crashed, so even a prior "finished" that
+// the plugin wrote should be corrected to "interrupted".
+//
+// "error" is left intact — a startup failure that wrote "error" directly (via
+// writeStartupError in sidecar.go) must not be overwritten to "interrupted" by
+// the pane-died hook, since "error" is the correct terminal state in that case.
+// "deleted" is still left intact — a deleted session should not be resurrected.
+// "interrupted" is also left alone to avoid a no-op double-write.
+//
+// Returns (true, nil) if the update was applied, (false, nil) if the row did
+// not exist, was already error/interrupted/deleted, or has ended_at set.
+func (d *DB) UpsertStatusInterruptedOverrideFinished(sessionName string) (bool, error) {
+	// Snapshot the current state before the write so that the advisory
+	// transition check (below) sees the from-state rather than the newly
+	// written to-state.
+	fromState, _ := d.currentStateOf(sessionName)
+	now := time.Now().UnixMilli()
+	const q = `
+UPDATE agent_status
+SET state = 'interrupted', last_seen = ?
+WHERE session_name = ?
+  AND ended_at IS NULL
+  AND state NOT IN ('error', 'interrupted', 'deleted')`
+	res, err := d.conn.Exec(q, now, sessionName)
+	if err != nil {
+		return false, fmt.Errorf("db: upsert status interrupted override finished: %w", err)
+	}
+	n, err := res.RowsAffected()
+	if err != nil {
+		return false, fmt.Errorf("db: upsert status interrupted override finished: rows affected: %w", err)
+	}
+	// Validate only when the write was actually applied (n > 0) and we have a
+	// prior state to validate from. When the SQL WHERE clause suppresses the
+	// write (session already interrupted or deleted), there is no transition
+	// to check.
+	if n > 0 && fromState != "" {
+		if terr := agent.Transition(fromState, agent.StateInterrupted); terr != nil {
+			fmt.Fprintf(os.Stderr, "[prism] UpsertStatusInterruptedOverrideFinished: invalid transition for session %q: %v\n",
+				sessionName, terr)
+		}
+	}
+	return n > 0, nil
+}
+
+// SetEnded marks the session as ended by setting ended_at to now.
+// It also cascades the ended_at to any child review-agent rows whose
+// session_name matches the pattern "<sessionName>~review-%", setting
+// ended_at only on rows where it is not already set (idempotent).
+// Both the parent and child updates are performed in a single transaction.
+//
+// The session name is escaped for SQL LIKE wildcards before being used as a
+// pattern prefix (using the same escaping as AllStatusesWithPrefix) so that
+// session names containing `%`, `_`, or `\` are handled correctly.
+func (d *DB) SetEnded(sessionName string) error {
+	now := time.Now().UnixMilli()
+	// Escape LIKE special characters in sessionName so that literal `%`, `_`,
+	// and `\` in the name are matched exactly, not as wildcards.
+	escaped := strings.NewReplacer(`\`, `\\`, `%`, `\%`, `_`, `\_`).Replace(sessionName)
+	tx, err := d.conn.Begin()
+	if err != nil {
+		return fmt.Errorf("db: set ended: begin tx: %w", err)
+	}
+	defer tx.Rollback() //nolint:errcheck
+
+	const q = `
+UPDATE agent_status
+SET    ended_at = ?
+WHERE  (session_name = ? OR session_name LIKE ? || '~review-%' ESCAPE '\')
+  AND  ended_at IS NULL`
+	if _, err := tx.Exec(q, now, sessionName, escaped); err != nil {
+		return fmt.Errorf("db: set ended: %w", err)
+	}
+
+	if err := tx.Commit(); err != nil {
+		return fmt.Errorf("db: set ended: commit: %w", err)
+	}
+	return nil
+}
+
+// MarkAllEnded marks every row in agent_status where ended_at IS NULL as ended
+// by setting ended_at = now (Unix milliseconds). The state column is intentionally
+// left unchanged — ended_at IS NULL is the canonical "active session" filter used
+// throughout the codebase; state captures the last known agent state before teardown
+// and is not overwritten here.
+//
+// It is used by `prism reset` to atomically close all live sessions in one
+// query rather than iterating over them individually.
+//
+// Returns the number of rows updated and any database error.
+// When there are no rows with ended_at IS NULL, returns (0, nil) — not an error.
+func (d *DB) MarkAllEnded() (int64, error) {
+	now := time.Now().UnixMilli()
+	res, err := d.conn.Exec(
+		"UPDATE agent_status SET ended_at = ? WHERE ended_at IS NULL",
+		now,
+	)
+	if err != nil {
+		return 0, fmt.Errorf("db: mark all ended: %w", err)
+	}
+	n, err := res.RowsAffected()
+	if err != nil {
+		return 0, fmt.Errorf("db: mark all ended: rows affected: %w", err)
+	}
+	return n, nil
+}
+
+// ClearEnded clears the ended_at timestamp for sessionName, making the session
+// visible again to AllActiveStatus and the dashboard (which both filter
+// WHERE ended_at IS NULL). Called when a session resumes from a terminal state
+// so that the resumed session re-appears in all active-session views.
+func (d *DB) ClearEnded(sessionName string) error {
+	_, err := d.conn.Exec(
+		"UPDATE agent_status SET ended_at = NULL WHERE session_name = ?",
+		sessionName,
+	)
+	if err != nil {
+		return fmt.Errorf("db: clear ended: %w", err)
+	}
+	return nil
+}
+
+// AllocatePort picks the lowest unused port from the range PortRangeStart–PortRangeEnd,
+// writes it to agent_status.harness_port for sessionName, and returns it.
+//
+// A port is considered "in use" if it is assigned to a session whose ended_at IS NULL
+// and harness_port IS NOT NULL. Ports assigned to ended sessions (ended_at IS NOT NULL)
+// are reclaimed and available for reuse.
+//
+// In addition to the DB check, each candidate port is probed at the OS level via
+// a brief TCP listen on 127.0.0.1:<port>. This prevents conflicts with non-prism
+// processes that happen to be using a port in the range.
+//
+// Returns an error if all ports in the range are exhausted or if the session does
+// not exist in agent_status.
+func (d *DB) AllocatePort(sessionName string) (int, error) {
+	// Collect ports currently assigned to active (non-ended) sessions.
+	rows, err := d.conn.Query(
+		"SELECT harness_port FROM agent_status WHERE ended_at IS NULL AND harness_port IS NOT NULL",
+	)
+	if err != nil {
+		return 0, fmt.Errorf("db: allocate port: query used ports: %w", err)
+	}
+	usedPorts := map[int]bool{}
+	for rows.Next() {
+		var p int
+		if err := rows.Scan(&p); err != nil {
+			rows.Close()
+			return 0, fmt.Errorf("db: allocate port: scan port: %w", err)
+		}
+		usedPorts[p] = true
+	}
+	rows.Close()
+	if err := rows.Err(); err != nil {
+		return 0, fmt.Errorf("db: allocate port: iterate ports: %w", err)
+	}
+
+	// Find the lowest unused port that is also available at the OS level.
+	for port := PortRangeStart; port <= PortRangeEnd; port++ {
+		if usedPorts[port] {
+			continue
+		}
+		if !portAvailable(port) {
+			continue
+		}
+		// Write the allocated port to agent_status.
+		res, err := d.conn.Exec(
+			"UPDATE agent_status SET harness_port = ? WHERE session_name = ?",
+			port, sessionName,
+		)
+		if err != nil {
+			return 0, fmt.Errorf("db: allocate port: update: %w", err)
+		}
+		n, err := res.RowsAffected()
+		if err != nil {
+			return 0, fmt.Errorf("db: allocate port: rows affected: %w", err)
+		}
+		if n == 0 {
+			return 0, fmt.Errorf("db: allocate port: session %q not found in agent_status", sessionName)
+		}
+		return port, nil
+	}
+
+	return 0, fmt.Errorf("db: allocate port: all ports in range %d–%d are exhausted", PortRangeStart, PortRangeEnd)
+}
+
+// ReleasePort sets harness_port = NULL for the given session.
+// Returns an error if the session does not exist in agent_status.
+// Calling ReleasePort on a session whose harness_port is already NULL is
+// idempotent and returns nil.
+func (d *DB) ReleasePort(sessionName string) error {
+	res, err := d.conn.Exec(
+		"UPDATE agent_status SET harness_port = NULL WHERE session_name = ?",
+		sessionName,
+	)
+	if err != nil {
+		return fmt.Errorf("db: release port: %w", err)
+	}
+	n, err := res.RowsAffected()
+	if err != nil {
+		return fmt.Errorf("db: release port: rows affected: %w", err)
+	}
+	if n == 0 {
+		return fmt.Errorf("db: release port: session %q not found in agent_status", sessionName)
+	}
+	return nil
+}
+
+// SetHostMode sets the host_mode column for the given session to 1 (true) or
+// 0 (false). Called by spawn when --host-mode is passed, so that cleanup can
+// skip container teardown for host-mode sessions.
+// It is a no-op when no row exists for sessionName (returns nil).
+func (d *DB) SetHostMode(sessionName string, hostMode bool) error {
+	val := 0
+	if hostMode {
+		val = 1
+	}
+	_, err := d.conn.Exec(
+		"UPDATE agent_status SET host_mode = ? WHERE session_name = ?",
+		val, sessionName,
+	)
+	if err != nil {
+		return fmt.Errorf("db: set host_mode: %w", err)
+	}
+	return nil
+}
+
+// SetIsolationMode records the resolved isolation mode for the given session.
+// mode is one of "podman", "bwrap", or "host". This is persisted so that
+// prism restore can re-spawn the session in the same isolation mode.
+// It is a no-op when no row exists for sessionName (returns nil).
+func (d *DB) SetIsolationMode(sessionName, mode string) error {
+	_, err := d.conn.Exec(
+		"UPDATE agent_status SET isolation_mode = ? WHERE session_name = ?",
+		mode, sessionName,
+	)
+	if err != nil {
+		return fmt.Errorf("db: set isolation_mode: %w", err)
+	}
+	return nil
+}
+
+// portAvailable checks whether a TCP port is available on localhost by
+// attempting a brief listen. Returns true if the port is free.
+func portAvailable(port int) bool {
+	ln, err := net.Listen("tcp", fmt.Sprintf("127.0.0.1:%d", port))
+	if err != nil {
+		return false
+	}
+	ln.Close()
+	return true
+}
+
+// RefreshWorktree unconditionally updates the repo and worktree columns for an
+// existing agent_status row. It also resets state to idle and refreshes
+// last_seen, making it useful for prism restore (which needs a clean idle state
+// regardless of what the prior session left behind). Unlike UpsertStatus, this
+// does not insert a new row when none exists — it is a no-op for unknown sessions.
+//
+// It is a no-op when no row exists for sessionName (returns nil).
+func (d *DB) RefreshWorktree(sessionName, repo, worktree string) error {
+	// RefreshWorktree is an administrative reset (correcting corrupted values
+	// during prism restore), not a normal lifecycle transition. It bypasses
+	// the state machine advisory check intentionally — idle is not a valid
+	// to-state in ValidTransitions, so calling checkTransition here would
+	// always produce spurious warnings on every restore invocation.
+	now := time.Now().UnixMilli()
+	_, err := d.conn.Exec(
+		`UPDATE agent_status
+		    SET repo = ?, worktree = ?, state = ?, last_seen = ?
+		  WHERE session_name = ?`,
+		repo, worktree, "idle", now, sessionName,
+	)
+	if err != nil {
+		return fmt.Errorf("db: refresh worktree: %w", err)
+	}
+	return nil
+}
+
+// SetInstanceID writes a UUID instance_id to the agent_status row for
+// sessionName. Called on tmux-session-start to uniquely identify this session
+// incarnation.
+func (d *DB) SetInstanceID(sessionName, instanceID string) error {
+	_, err := d.conn.Exec(
+		"UPDATE agent_status SET instance_id = ? WHERE session_name = ?",
+		instanceID, sessionName,
+	)
+	if err != nil {
+		return fmt.Errorf("db: set instance_id: %w", err)
+	}
+	return nil
+}
+
+// SetGroupID writes a group_id to the agent_status row for sessionName. Called
+// by SpawnSession when opts.GroupID is non-empty to associate the new session
+// with a session_groups entry (Issue E hook — see #849 §3.1 and #860).
+//
+// No-op when sessionName has no agent_status row (the UPDATE affects zero rows
+// and returns nil).
+func (d *DB) SetGroupID(sessionName, groupID string) error {
+	_, err := d.conn.Exec(
+		"UPDATE agent_status SET group_id = ? WHERE session_name = ?",
+		groupID, sessionName,
+	)
+	if err != nil {
+		return fmt.Errorf("db: set group_id: %w", err)
+	}
+	return nil
+}
+
+// ClearInstanceID sets instance_id to NULL for sessionName. Called on
+// tmux-session-end to mark the session incarnation as over.
+func (d *DB) ClearInstanceID(sessionName string) error {
+	_, err := d.conn.Exec(
+		"UPDATE agent_status SET instance_id = NULL WHERE session_name = ?",
+		sessionName,
+	)
+	if err != nil {
+		return fmt.Errorf("db: clear instance_id: %w", err)
+	}
+	return nil
+}
+
+// UpdateHarnessSessionID unconditionally sets harness_session_id for sessionName
+// to the given sid value. Unlike UpsertStatus (which only updates harness_session_id
+// via COALESCE when non-nil), this always overwrites — allowing the sidecar to
+// keep the stored SID current when the user creates a new harness session
+// mid-conversation (e.g. via /continue or TUI restart).
+//
+// It writes to both agent_status and sessions (matching by instance_id) so
+// that cleanup.runSessionArchive can read harness_session_id directly from the
+// sessions row without a fallback query. The sessions update is best-effort:
+// if no row exists for the current instance_id it is silently skipped.
+//
+// It is a no-op when no row exists for sessionName (returns nil).
+func (d *DB) UpdateHarnessSessionID(sessionName, sid string) error {
+	tx, err := d.conn.Begin()
+	if err != nil {
+		return fmt.Errorf("db: update harness_session_id: begin tx: %w", err)
+	}
+	defer tx.Rollback() //nolint:errcheck
+
+	if _, err := tx.Exec(
+		"UPDATE agent_status SET harness_session_id = ? WHERE session_name = ?",
+		sid, sessionName,
+	); err != nil {
+		return fmt.Errorf("db: update harness_session_id (agent_status): %w", err)
+	}
+
+	// Also update sessions.harness_session_id for the current incarnation so
+	// that cleanup reads the correct value without a fallback query.
+	if _, err := tx.Exec(`
+		UPDATE sessions
+		   SET harness_session_id = ?
+		 WHERE instance_id = (
+		         SELECT instance_id FROM agent_status WHERE session_name = ?
+		       )`, sid, sessionName,
+	); err != nil {
+		return fmt.Errorf("db: update harness_session_id (sessions): %w", err)
+	}
+
+	if err := tx.Commit(); err != nil {
+		return fmt.Errorf("db: update harness_session_id: commit: %w", err)
+	}
+	return nil
+}
+
+// HarnessSessionIDForInstance returns the harness_session_id stored in
+// agent_status for the given instance_id, or "" when not found or NULL.
+// Used by cleanup as a fallback when sessions.harness_session_id is NULL
+// (e.g. for sessions started before UpdateHarnessSessionID was fixed to
+// also write to sessions).
+func (d *DB) HarnessSessionIDForInstance(instanceID string) (string, error) {
+	var sid sql.NullString
+	err := d.conn.QueryRow(
+		"SELECT harness_session_id FROM agent_status WHERE instance_id = ?",
+		instanceID,
+	).Scan(&sid)
+	if errors.Is(err, sql.ErrNoRows) {
+		return "", nil
+	}
+	if err != nil {
+		return "", fmt.Errorf("db: harness_session_id for instance: %w", err)
+	}
+	if !sid.Valid {
+		return "", nil
+	}
+	return sid.String, nil
+}
+
+// CurrentStatus returns the agent_status row for sessionName, or nil if not found.
+func (d *DB) CurrentStatus(sessionName string) (*Status, error) {
+	const q = `
+SELECT session_name, repo, worktree, state, title, agent_name, model_id, root_agent_name, root_model_id, host_mode, isolation_mode, instance_id, last_seen, ended_at, harness, harness_session_id, harness_port, group_id
+FROM agent_status
+WHERE session_name = ?`
+	row := d.conn.QueryRow(q, sessionName)
+	s, err := scanStatus(row)
+	if err == sql.ErrNoRows {
+		return nil, nil
+	}
+	if err != nil {
+		return nil, fmt.Errorf("db: current status: %w", err)
+	}
+	return s, nil
+}
+
+// AllActiveStatus returns all agent_status rows where ended_at IS NULL.
+func (d *DB) AllActiveStatus() ([]Status, error) {
+	const q = `
+SELECT session_name, repo, worktree, state, title, agent_name, model_id, root_agent_name, root_model_id, host_mode, isolation_mode, instance_id, last_seen, ended_at, harness, harness_session_id, harness_port, group_id
+FROM agent_status
+WHERE ended_at IS NULL`
+	return d.queryStatuses(q)
+}
+
+// ActiveSessionCountForMode returns the number of agent_status rows where
+// ended_at IS NULL AND isolation_mode = mode. Used by the per-isolator
+// concurrency cap checks. Returns 0 when no rows match.
+//
+// Replaces the mode-specific ActiveBwrapSessionCount and
+// ActiveSandboxExecSessionCount helpers; callable for any IsolationMode value.
+func (d *DB) ActiveSessionCountForMode(mode string) (int, error) {
+	var n int
+	err := d.conn.QueryRow(
+		"SELECT COUNT(*) FROM agent_status WHERE ended_at IS NULL AND isolation_mode = ?",
+		mode,
+	).Scan(&n)
+	if err != nil {
+		return 0, fmt.Errorf("db: active session count for mode %q: %w", mode, err)
+	}
+	return n, nil
+}
+
+// ActiveSessionsForMode returns the agent_status rows where ended_at IS NULL
+// AND isolation_mode = mode, suitable for cap-error detail listings.
+//
+// Replaces the mode-specific ActiveBwrapSessions and ActiveSandboxExecSessions
+// helpers; callable for any IsolationMode value.
+func (d *DB) ActiveSessionsForMode(mode string) ([]Status, error) {
+	const q = `
+SELECT session_name, repo, worktree, state, title, agent_name, model_id, root_agent_name, root_model_id, host_mode, isolation_mode, instance_id, last_seen, ended_at, harness, harness_session_id, harness_port, group_id
+FROM agent_status
+WHERE ended_at IS NULL AND isolation_mode = ?`
+	return d.queryStatuses(q, mode)
+}
+
+// AllActiveStatusForRepo returns all active agent_status rows for repo.
+func (d *DB) AllActiveStatusForRepo(repo string) ([]Status, error) {
+	const q = `
+SELECT session_name, repo, worktree, state, title, agent_name, model_id, root_agent_name, root_model_id, host_mode, isolation_mode, instance_id, last_seen, ended_at, harness, harness_session_id, harness_port, group_id
+FROM agent_status
+WHERE ended_at IS NULL AND repo = ?`
+	return d.queryStatuses(q, repo)
+}
+
+// AllStatusesWithPrefix returns all agent_status rows (active and ended)
+// whose session_name starts with the given prefix. Used by `prism checkin
+// <parent>~review` to enumerate all review rounds including completed ones.
+//
+// The prefix is matched using LIKE with proper escaping of SQL LIKE wildcard
+// characters (`%`, `_`, `\`) so that session names containing these characters
+// are handled correctly.
+func (d *DB) AllStatusesWithPrefix(prefix string) ([]Status, error) {
+	// Escape LIKE special characters in the prefix so literal underscores and
+	// percent signs in session names are matched exactly, not as wildcards.
+	escaped := strings.NewReplacer(`\`, `\\`, `%`, `\%`, `_`, `\_`).Replace(prefix)
+	const q = `
+SELECT session_name, repo, worktree, state, title, agent_name, model_id, root_agent_name, root_model_id, host_mode, isolation_mode, instance_id, last_seen, ended_at, harness, harness_session_id, harness_port, group_id
+FROM agent_status
+WHERE session_name LIKE ? ESCAPE '\'`
+	return d.queryStatuses(q, escaped+"%")
+}
+
+// WaitingCount returns the number of active sessions with state='waiting'.
+func (d *DB) WaitingCount() (int, error) {
+	var n int
+	err := d.conn.QueryRow(
+		"SELECT COUNT(*) FROM agent_status WHERE state = 'waiting' AND ended_at IS NULL",
+	).Scan(&n)
+	if err != nil {
+		return 0, fmt.Errorf("db: waiting count: %w", err)
+	}
+	return n, nil
+}
+
+// CoordinatorForRepo returns the agent_status row for the active coordinator
+// session of repo (i.e. the row where repo = repo AND root_agent_name =
+// "coordinator" AND ended_at IS NULL). Returns nil when no coordinator exists.
+// When multiple rows match (schema violation), the most-recently-seen row is
+// returned and a duplicate is silently tolerated.
+func (d *DB) CoordinatorForRepo(repo string) (*Status, error) {
+	const q = `
+SELECT session_name, repo, worktree, state, title, agent_name, model_id, root_agent_name, root_model_id, host_mode, isolation_mode, instance_id, last_seen, ended_at, harness, harness_session_id, harness_port, group_id
+FROM agent_status
+WHERE repo = ? AND root_agent_name = 'coordinator' AND ended_at IS NULL
+ORDER BY last_seen DESC
+LIMIT 1`
+	row := d.conn.QueryRow(q, repo)
+	s, err := scanStatus(row)
+	if err == sql.ErrNoRows {
+		return nil, nil
+	}
+	if err != nil {
+		return nil, fmt.Errorf("db: coordinator for repo: %w", err)
+	}
+	return s, nil
+}
+
+// RootAgentName returns the root_agent_name for sessionName, or "" when the
+// row does not exist or root_agent_name is NULL (pre-migration row).
+// The second return value (rowExists) distinguishes the two empty cases:
+//   - ("", false, nil)  — no agent_status row found (new/unknown session)
+//   - ("", true, nil)   — row found but root_agent_name is NULL (pre-migration)
+//   - (name, true, nil) — row found with a populated root_agent_name
+func (d *DB) RootAgentName(sessionName string) (name string, rowExists bool, err error) {
+	var ns sql.NullString
+	const q = `SELECT root_agent_name FROM agent_status WHERE session_name = ?`
+	if scanErr := d.conn.QueryRow(q, sessionName).Scan(&ns); scanErr != nil {
+		if scanErr == sql.ErrNoRows {
+			return "", false, nil
+		}
+		return "", false, fmt.Errorf("db: root agent name: %w", scanErr)
+	}
+	if !ns.Valid {
+		return "", true, nil
+	}
+	return ns.String, true, nil
+}
+
+// queryStatuses is a helper that runs a SELECT on agent_status and scans rows.
+func (d *DB) queryStatuses(q string, args ...any) ([]Status, error) {
+	rows, err := d.conn.Query(q, args...)
+	if err != nil {
+		return nil, fmt.Errorf("db: query statuses: %w", err)
+	}
+	defer rows.Close()
+
+	var statuses []Status
+	for rows.Next() {
+		s, err := scanStatus(rows)
+		if err != nil {
+			return nil, fmt.Errorf("db: scan status: %w", err)
+		}
+		statuses = append(statuses, *s)
+	}
+	if err := rows.Err(); err != nil {
+		return nil, fmt.Errorf("db: iterate statuses: %w", err)
+	}
+	return statuses, nil
+}
+
+// scanner abstracts *sql.Row and *sql.Rows for scanStatus.
+type scanner interface {
+	Scan(dest ...any) error
+}
+
+func scanStatus(s scanner) (*Status, error) {
+	var st Status
+	var lastSeen int64
+	var endedAt sql.NullInt64
+	var hostMode sql.NullInt64
+	var instanceID sql.NullString
+	var harness sql.NullString
+	var harnessSessionID sql.NullString
+	var harnessPort sql.NullInt64
+	var isolationMode sql.NullString
+	var groupID sql.NullString
+	err := s.Scan(
+		&st.SessionName, &st.Repo, &st.Worktree, &st.State,
+		&st.Title, &st.AgentName, &st.ModelID,
+		&st.RootAgentName, &st.RootModelID, &hostMode, &isolationMode, &instanceID, &lastSeen, &endedAt,
+		&harness, &harnessSessionID, &harnessPort, &groupID,
+	)
+	if err != nil {
+		return nil, err
+	}
+	if groupID.Valid {
+		g := groupID.String
+		st.GroupID = &g
+	}
+	st.LastSeen = time.UnixMilli(lastSeen)
+	if endedAt.Valid {
+		t := time.UnixMilli(endedAt.Int64)
+		st.EndedAt = &t
+	}
+	// host_mode: treat NULL (rows written before migration) as 0/false.
+	if hostMode.Valid {
+		st.HostMode = hostMode.Int64 != 0
+	}
+	// isolation_mode: NULL means not recorded (pre-v10 row); callers derive
+	// the mode from host_mode for back-compat in that case.
+	if isolationMode.Valid {
+		st.IsolationMode = isolationMode.String
+	}
+	if instanceID.Valid {
+		id := instanceID.String
+		st.InstanceID = &id
+	}
+	if harness.Valid {
+		h := harness.String
+		st.Harness = &h
+	}
+	if harnessSessionID.Valid {
+		hsid := harnessSessionID.String
+		st.HarnessSessionID = &hsid
+	}
+	if harnessPort.Valid {
+		hp := int(harnessPort.Int64)
+		st.HarnessPort = &hp
+	}
+	return &st, nil
+}
